Fix Filesystem::dataPoolSize's return
It was returning 0 by default when the given pool was not found.
It now returns -ENOENT instead and also uses ssize_t instead of int
for the return type.

+++ src/Filesystem.hh
@@ -80,7 +80,7 @@ public:
 
   std::string dataPoolPrefix(const std::string &pool) const;
 
-  int dataPoolSize(const std::string &pool) const;
+  ssize_t dataPoolSize(const std::string &pool) const;
 
   int addMetadataPool(const std::string &name, const std::string &prefix);
